Check for _site before removing

+++ .eleventy.js
@@ -2,7 +2,9 @@ const fs = require("fs");
 
 module.exports = function (eleventyConfig) {
   // clean output directory
-  fs.rmSync("_site", { recursive: true });
+  if (fs.existsSync("_site")) {
+    fs.rmSync("_site", { recursive: true });
+  }
 
   eleventyConfig.setTemplateFormats(["njk", "png", "md", "svg"]);
   eleventyConfig.addPassthroughCopy("src/css");
